Use size_t for sizes in CFX_BitStream.
-- beef up some tests.

+++ core/fxcrt/cfx_bitstream.h
@@ -20,19 +20,19 @@ class CFX_BitStream {
   void ByteAlign();
 
   bool IsEOF() const { return m_BitPos >= m_BitSize; }
-  uint32_t GetPos() const { return m_BitPos; }
+  size_t GetPos() const { return m_BitPos; }
   uint32_t GetBits(uint32_t nBits);
 
-  void SkipBits(uint32_t nBits) { m_BitPos += nBits; }
+  void SkipBits(size_t nBits) { m_BitPos += nBits; }
   void Rewind() { m_BitPos = 0; }
 
-  uint32_t BitsRemaining() const {
+  size_t BitsRemaining() const {
     return m_BitSize >= m_BitPos ? m_BitSize - m_BitPos : 0;
   }
 
  private:
-  uint32_t m_BitPos = 0;
-  const uint32_t m_BitSize;
+  size_t m_BitPos = 0;
+  const size_t m_BitSize;
   UnownedPtr<const uint8_t> const m_pData;
 };
